requirements.txt
httpx>=0.27
google-genai>=0.8
schedule>=1.2
python-dotenv>=1.0
streamlit>=1.35
plotly>=5.20
